refactor(eve): keep Vercel command adapter in place
