+++ packages/eve/src/execution/sandbox/bindings/vercel.ts
@@ -42,11 +42,11 @@ import {
 import { getNamedVercelSandbox } from "#execution/sandbox/bindings/vercel-lookup.js";
 import { normalizeVercelReadStream } from "#execution/sandbox/bindings/vercel-read-stream.js";
 import type {
+  VercelCommand,
   VercelCreateOptions,
   VercelModule,
   VercelSandbox,
 } from "#execution/sandbox/bindings/vercel-sdk-types.js";
-import { adaptVercelCommandToSandboxProcess } from "#execution/sandbox/bindings/vercel-command.js";
 
 export interface CreateVercelSandboxInput {
   readonly createSandbox?: CreateVercelSandbox;
@@ -408,20 +408,15 @@ function createSessionCreateParams(
   }
 
   /*
-   * Strip both `source` and `runtime` from author-supplied create
-   * options for the template-backed session path. The framework owns
-   * the session source there, and the Vercel SDK rejects `runtime`
-   * when `source` is a snapshot because the runtime is already baked
-   * into the snapshot filesystem.
+   * Strip both `source` and `runtime` from author-supplied create options for the
+   * template-backed session path. The framework owns the source there, and the SDK
+   * rejects `runtime` for a snapshot because its runtime is already baked into it.
    */
   const {
     runtime: _runtime,
     source: _source,
     ...sessionCreateOptions
-  } = input.createOptions as VercelCreateOptions & {
-    readonly runtime?: unknown;
-    readonly source?: unknown;
-  };
+  } = input.createOptions as VercelCreateOptions & Partial<Record<"runtime" | "source", unknown>>;
 
   return {
     ...sessionCreateOptions,
@@ -505,6 +500,72 @@ function createVercelInternalSandboxSession(
         recursive: options.recursive,
         signal: options.abortSignal,
       });
+    },
+  };
+}
+
+/**
+ * Wraps a Vercel `Command` (returned from `runCommand({ detached: true })`)
+ * in the AI SDK `Experimental_SandboxProcess` shape. Splits the single
+ * `logs()` iterator into two byte streams, exposes a `wait()` that
+ * resolves with the exit code, and forwards `kill()` to the SDK.
+ */
+function adaptVercelCommandToSandboxProcess(command: VercelCommand): SandboxProcess {
+  const encoder = new TextEncoder();
+  let stdoutController: ReadableStreamDefaultController<Uint8Array> | undefined;
+  let stderrController: ReadableStreamDefaultController<Uint8Array> | undefined;
+  let streamingDone = false;
+  let streamingError: unknown;
+
+  const stdout = new ReadableStream<Uint8Array>({
+    start(controller) {
+      stdoutController = controller;
+    },
+  });
+  const stderr = new ReadableStream<Uint8Array>({
+    start(controller) {
+      stderrController = controller;
+    },
+  });
+
+  void (async () => {
+    try {
+      for await (const message of command.logs()) {
+        const chunk = encoder.encode(message.data);
+        if (message.stream === "stdout") {
+          stdoutController?.enqueue(chunk);
+        } else {
+          stderrController?.enqueue(chunk);
+        }
+      }
+    } catch (error) {
+      streamingError = error;
+      stdoutController?.error(error);
+      stderrController?.error(error);
+    } finally {
+      streamingDone = true;
+      if (streamingError === undefined) {
+        stdoutController?.close();
+        stderrController?.close();
+      }
+    }
+  })();
+
+  return {
+    stdout,
+    stderr,
+    async wait() {
+      const finished = await command.wait();
+      while (!streamingDone) {
+        await new Promise((resolve) => setTimeout(resolve, 0));
+      }
+      if (streamingError !== undefined) {
+        throw streamingError;
+      }
+      return { exitCode: finished.exitCode };
+    },
+    async kill() {
+      await command.kill();
     },
   };
 }
